% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder2-package.R
\docType{package}
\name{gapminder2-package}
\alias{gapminder2}
\alias{gapminder2-package}
\title{gapminder2: Additional data from gapminder}
\description{
An excerpt of the data available at Gapminder.org. For each of 133
    countries, the package provides values for life expectancy, GDP per capita,
    and population, and mean school in years for men aged 25-34 years old, from 1972 to 2007.
}
\author{
\strong{Maintainer}: Seevasant Indran \email{zeeva85@gmail.com}

}
\keyword{internal}
