% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapminder2.R
\docType{data}
\name{gapminder2}
\alias{gapminder2}
\title{Data from gapminder}
\format{A data frame with 1064 rows and 7 variables:
\describe{
  \item{country}{countries in the gapminder dataset}
  \item{continent}{continent the country belongs}
  \item{year}{year the data was acquired}
  \item{lifeExp}{mean life epectancy}
  \item{gdpPercap}{GDP per capita}
  \item{meanSchool}{mean years in school men 25 -35 years old}
}}
\usage{
gapminder2
}
\description{
Data from gapminder
}
\keyword{datasets}
