% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_country.R
\name{subset_country}
\alias{subset_country}
\title{Subset a country using a set of inputs}
\usage{
subset_country()
}
\arguments{
\item{input}{the country or countries without "" in the console.}
}
\value{
A list of country that is subseted from the gapminder2 dataset:
}
\description{
of country names to subset a the gapminder2 dataset
reciprocal/inverse of a vector.
}
\details{
Provide a list in the console to subset the gapminder dataset more interactively.
}
\examples{
subset_country()

Console preview:

> subset_country()
1: Japan
2: Malaysia
3:

}
